+++ src/test/clojure/com/github/clojure_repl/intellij/repl_command_test.clj
@@ -7,12 +7,12 @@
   (testing "clojure"
     (is (= ["clojure"
             "-Sdeps"
-            "{:deps {nrepl/nrepl {:mvn/version \"1.0.0\"} cider/cider-nrepl {:mvn/version \"0.45.0\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
+            "{:deps {nrepl/nrepl {:mvn/version \"1.3.1\"} cider/cider-nrepl {:mvn/version \"0.55.7\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
             "-M:cider/nrepl"]
            (repl-command/project->repl-start-command :clojure [] [])))
     (is (= ["clojure"
             "-Sdeps"
-            "{:deps {nrepl/nrepl {:mvn/version \"1.0.0\"} cider/cider-nrepl {:mvn/version \"0.45.0\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
+            "{:deps {nrepl/nrepl {:mvn/version \"1.3.1\"} cider/cider-nrepl {:mvn/version \"0.55.7\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
             "-M:foo-b:bar:cider/nrepl"]
            (repl-command/project->repl-start-command :clojure ["foo-b" "bar"] [])))
     (testing "windows powershell"
@@ -23,7 +23,7 @@
                     repl-command/shell (constantly {:exit 0})]
         (is (= ["/full/path/powershell" "-NoProfile" "-Command" "clojure"
                 "-Sdeps"
-                "{:deps {nrepl/nrepl {:mvn/version \"1.0.0\"} cider/cider-nrepl {:mvn/version \"0.45.0\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
+                "{:deps {nrepl/nrepl {:mvn/version \"1.3.1\"} cider/cider-nrepl {:mvn/version \"0.55.7\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
                 "-M:cider/nrepl"]
                (repl-command/project->repl-start-command :clojure [] [])))))
     (testing "windows pwsh"
@@ -34,16 +34,16 @@
                     repl-command/shell (constantly {:exit 0})]
         (is (= ["/full/path/pwsh" "-NoProfile" "-Command" "clojure"
                 "-Sdeps"
-                "{:deps {nrepl/nrepl {:mvn/version \"1.0.0\"} cider/cider-nrepl {:mvn/version \"0.45.0\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
+                "{:deps {nrepl/nrepl {:mvn/version \"1.3.1\"} cider/cider-nrepl {:mvn/version \"0.55.7\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
                 "-M:cider/nrepl"]
                (repl-command/project->repl-start-command :clojure [] []))))))
   (testing "lein"
-    (is (= ["lein" "update-in" ":dependencies" "conj" "[nrepl/nrepl \"1.0.0\"]"
-            "--" "update-in" ":plugins" "conj" "[cider/cider-nrepl \"0.45.0\"]"
+    (is (= ["lein" "update-in" ":dependencies" "conj" "[nrepl/nrepl \"1.3.1\"]"
+            "--" "update-in" ":plugins" "conj" "[cider/cider-nrepl \"0.55.7\"]"
             "--" "repl" ":headless" ":host" "localhost"]
            (repl-command/project->repl-start-command :lein [] [])))
-    (is (= ["lein" "update-in" ":dependencies" "conj" "[nrepl/nrepl \"1.0.0\"]"
-            "--" "update-in" ":plugins" "conj" "[cider/cider-nrepl \"0.45.0\"]"
+    (is (= ["lein" "update-in" ":dependencies" "conj" "[nrepl/nrepl \"1.3.1\"]"
+            "--" "update-in" ":plugins" "conj" "[cider/cider-nrepl \"0.55.7\"]"
             "--" "with-profile" "+foo-b,+bar" "repl" ":headless" ":host" "localhost"]
            (repl-command/project->repl-start-command :lein ["foo-b" "bar"] [])))
     (testing "windows powershell"
@@ -53,8 +53,8 @@
                                                        "/full/path/powershell"))
                     repl-command/shell (constantly {:exit 0})]
         (is (= ["/full/path/powershell" "-NoProfile" "-Command" "lein"
-                "update-in" ":dependencies" "conj" "[nrepl/nrepl \"1.0.0\"]"
-                "--" "update-in" ":plugins" "conj" "[cider/cider-nrepl \"0.45.0\"]"
+                "update-in" ":dependencies" "conj" "[nrepl/nrepl \"1.3.1\"]"
+                "--" "update-in" ":plugins" "conj" "[cider/cider-nrepl \"0.55.7\"]"
                 "--" "repl" ":headless" ":host" "localhost"]
                (repl-command/project->repl-start-command :lein [] [])))))
     (testing "windows pwsh"
@@ -64,8 +64,8 @@
                                                        "/full/path/pwsh"))
                     repl-command/shell (constantly {:exit 0})]
         (is (= ["/full/path/pwsh" "-NoProfile" "-Command" "lein"
-                "update-in" ":dependencies" "conj" "[nrepl/nrepl \"1.0.0\"]"
-                "--" "update-in" ":plugins" "conj" "[cider/cider-nrepl \"0.45.0\"]"
+                "update-in" ":dependencies" "conj" "[nrepl/nrepl \"1.3.1\"]"
+                "--" "update-in" ":plugins" "conj" "[cider/cider-nrepl \"0.55.7\"]"
                 "--" "repl" ":headless" ":host" "localhost"]
                (repl-command/project->repl-start-command :lein [] []))))))
   (testing "babashka"
@@ -82,7 +82,7 @@
            (repl-command/project->repl-start-command :nbb [] []))))
   (testing "gradle"
     (is (= ["./gradlew"
-            "-Pdev.clojurephant.jack-in.nrepl=nrepl:nrepl:1.0.0,cider:cider-nrepl:0.45.0"
+            "-Pdev.clojurephant.jack-in.nrepl=nrepl:nrepl:1.3.1,cider:cider-nrepl:0.55.7"
             "clojureRepl"
             "--middleware=cider.nrepl/cider-middleware"]
            (repl-command/project->repl-start-command :gradle [] [])))))