+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add client info to clone op for metrics.
+- Bump nrepl + cider-nrepl version on local repl.
 
 ## 2.5.2